Add labels to recipe list and changed recipe list style
