+++ components/index.ts
@@ -1,5 +1,6 @@
-export * from "./RecipeList";
+export * from "./RecipeList/RecipeList";
 export * from "./PageLayout";
 export * from "./RecipeSearch";
 export * from "./MealPlanRecipeItem";
-export * from "./MealPlanRecipeList"
+export * from "./MealPlanRecipeList";
+export * from "./RecipeList";
